be/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cryptocode==0.1
dnspython==2.7.0
email_validator==2.1.1
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.11
packaging==24.2
pip-tools==7.4.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
pydantic==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.2.0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
spotipy==2.24.0
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.1.0
typer==0.13.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
